Add new URL, view and template to display lists
Move complexity from home page to a new list page template.
Now home page only shows start a new list, and once selected
redirects the user to a new URL where list entries can be
added and displayed.  This is a step towards a unique URL
and to-do list for each user